# observability: constant time implementation of prometheusFieldName